check length of each parsed length
some lines can be formatted in such a way to mess up parser, so added check to make sure output row is right length. If it isn't that line/record will be written to file so you can see what's going on instead of writing to CSV file where you might miss it.